# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIL Backend Assessment API",
    "description": "Comprehensive REST API for managing users, customers, categories, products, and orders with JWT and OIDC authentication support.",
    "version": "1.0.0",
    "contact": {
      "name": "SIL Backend Assessment",
      "url": "https://github.com/sil-backend-assessment"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Use JWT token for user operations or OIDC token for customer operations."
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the order"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "description": "Customer ID who placed the order"
          },
          "order_number": {
            "type": "string",
            "description": "Unique order number"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "CONFIRMED", "PROCESSING", "SHIPPED", "DELIVERED", "CANCELLED"],
            "description": "Order status"
          },
          "total_amount": {
            "type": "number",
            "format": "float",
            "description": "Total order amount"
          },
          "shipping_address": {
            "type": "string",
            "description": "Shipping address"
          },
          "billing_address": {
            "type": "string",
            "description": "Billing address"
          },
          "notes": {
            "type": "string",
            "description": "Order notes"
          },
          "order_date": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the order was placed"
          },
          "shipped_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Date when the order was shipped"
          },
          "delivered_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Date when the order was delivered"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the order was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the order was last updated"
          }
        },
        "required": ["id", "customer_id", "order_number", "status", "total_amount", "shipping_address", "billing_address", "order_date", "created_at", "updated_at"]
      },
      "CreateOrderRequest": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "description": "Customer ID who is placing the order"
          },
          "shipping_address": {
            "type": "string",
            "description": "Shipping address"
          },
          "billing_address": {
            "type": "string",
            "description": "Billing address"
          },
          "notes": {
            "type": "string",
            "description": "Order notes"
          },
          "order_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemRequest"
            },
            "description": "List of order items"
          }
        },
        "required": ["customer_id", "shipping_address", "billing_address", "order_items"]
      },
      "CreateOrderItemRequest": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "description": "Product ID"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Quantity of the product"
          }
        },
        "required": ["product_id", "quantity"]
      },
      "UpdateOrderRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["PENDING", "CONFIRMED", "PROCESSING", "SHIPPED", "DELIVERED", "CANCELLED"],
            "description": "Order status"
          },
          "shipping_address": {
            "type": "string",
            "description": "Shipping address"
          },
          "billing_address": {
            "type": "string",
            "description": "Billing address"
          },
          "notes": {
            "type": "string",
            "description": "Order notes"
          },
          "shipped_date": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the order was shipped"
          },
          "delivered_date": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the order was delivered"
          }
        }
      },
      "CustomerProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the customer"
          },
          "first_name": {
            "type": "string",
            "description": "Customer's first name"
          },
          "last_name": {
            "type": "string",
            "description": "Customer's last name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Customer's email address"
          },
          "phone": {
            "type": "string",
            "description": "Customer's phone number"
          },
          "address": {
            "type": "string",
            "description": "Customer's address"
          },
          "city": {
            "type": "string",
            "description": "Customer's city"
          },
          "state": {
            "type": "string",
            "description": "Customer's state"
          },
          "zip_code": {
            "type": "string",
            "description": "Customer's zip code"
          },
          "country": {
            "type": "string",
            "description": "Customer's country"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the customer was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the customer was last updated"
          }
        },
        "required": ["id", "first_name", "last_name", "email", "created_at", "updated_at"]
      },
...
      "OIDCAuthURLResponse": {
        "type": "object",
        "properties": {
          "auth_url": {
            "type": "string",
            "format": "uri",
            "description": "OIDC authorization URL"
          },
          "state": {
            "type": "string",
            "description": "State parameter for CSRF protection"
          }
        },
        "required": ["auth_url", "state"]
      },
      "OIDCLoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT access token"
          },
          "refresh_token": {
            "type": "string",
            "description": "JWT refresh token"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerProfile"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Token expiration time"
          },
          "is_new_user": {
            "type": "boolean",
            "description": "Whether this is a new user registration"
          }
        },
        "required": ["access_token", "refresh_token", "customer", "expires_at", "is_new_user"]
      },
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message"
          }
        },
        "required": ["message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
      "PaginatedOrdersResponse": {
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page"
          },
          "offset": {
            "type": "integer",
            "description": "Number of items skipped"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Filter by customer ID"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "CONFIRMED", "PROCESSING", "SHIPPED", "DELIVERED", "CANCELLED"],
            "nullable": true,
            "description": "Filter by order status"
          }
        },
        "required": ["orders", "limit", "offset"]
      }
    }
  },
  "paths": {
...
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get all orders",
        "description": "Retrieve all orders with pagination and filtering",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of orders to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of orders to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "customer_id",
            "in": "query",
            "description": "Filter by customer ID",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by order status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["PENDING", "CONFIRMED", "PROCESSING", "SHIPPED", "DELIVERED", "CANCELLED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrdersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create a new order",
        "description": "Create a new order with the provided information",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID",
        "description": "Retrieve a specific order by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Order ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Orders"],
        "summary": "Update order",
        "description": "Update an existing order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Order ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Orders"],
        "summary": "Delete order",
        "description": "Delete an existing order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Order ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oidc/login": {
      "get": {
        "tags": ["OIDC Authentication"],
        "summary": "Get OIDC authorization URL",
        "description": "Generate OIDC authorization URL for customer login",
        "security": [],
        "responses": {
          "200": {
            "description": "Authorization URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OIDCAuthURLResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oidc/callback": {
      "get": {
        "tags": ["OIDC Authentication"],
        "summary": "Handle OIDC callback",
        "description": "Handle OIDC provider callback and authenticate customer",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "description": "Authorization code from OIDC provider",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "description": "State parameter for CSRF protection",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OIDCLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oidc/validate": {
      "get": {
        "tags": ["OIDC Authentication"],
        "summary": "Validate OIDC token",
        "description": "Validate an OIDC token and return user information",
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sub": {
                      "type": "string",
                      "description": "Subject identifier"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "description": "User email"
                    },
                    "name": {
                      "type": "string",
                      "description": "User full name"
                    },
                    "given_name": {
                      "type": "string",
                      "description": "User first name"
                    },
                    "family_name": {
                      "type": "string",
                      "description": "User last name"
                    },
                    "picture": {
                      "type": "string",
                      "format": "uri",
                      "description": "User profile picture URL"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}